Update the migrations

+++ src/database/migrations/2018_09_01_165420_create_employees_table.php
@@ -23,28 +23,33 @@ class CreateEmployeesTable extends Migration
      */
     public function up()
     {
-        Schema::create($this->tableName, function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('gender_id')->nullable();
-            $table->unsignedInteger('title_id')->nullable();
-            $table->unsignedInteger('manager_id')->nullable();
-            $table->string('country_id')->nullable();
-            $table->string('first_name');
-            $table->string('middle_names')->nullable();
-            $table->string('last_name');
-            $table->date('date_of_birth');
-            $table->string('national_insurance_no');
-            $table->string('address_line_1');
-            $table->string('address_line_2');
-            $table->string('city');
-            $table->string('county_state');
-            $table->timestamps();
-            $table->softDeletes();
+        if (! Schema::hasTable($this->tableName)) {
+            Schema::create($this->tableName, function (Blueprint $table) {
+                $table->increments('id');
+                $table->unsignedInteger('gender_id')->nullable();
+                $table->unsignedInteger('title_id')->nullable();
+                $table->unsignedInteger('manager_id')->nullable();
+                $table->string('country_id')->nullable();
+                $table->string('first_name');
+                $table->string('middle_names')->nullable();
+                $table->string('last_name');
+                $table->date('date_of_birth');
+                $table->string('national_insurance_no');
+                $table->string('address_line_1');
+                $table->string('address_line_2');
+                $table->string('city');
+                $table->string('county_state');
+                $table->timestamps();
+                $table->softDeletes();
 
-            $table->foreign('gender_id')->references('id')->on('genders')->onDelete('set null')->onUpdate('cascade');
-            $table->foreign('title_id')->references('id')->on('titles')->onDelete('set null')->onUpdate('cascade');
-            $table->foreign('manager_id')->references('id')->on($this->tableName)->onDelete('set null')->onUpdate('cascade');
-        });
+                $table->foreign('gender_id')->references('id')->on('genders')->onDelete('set null')->onUpdate('cascade');
+                $table->foreign('title_id')->references('id')->on('titles')->onDelete('set null')->onUpdate('cascade');
+                $table->foreign('manager_id')->references('id')->on($this->tableName)->onDelete('set null')->onUpdate('cascade');
+            });
+        } else {
+            console_write($this->tableName . ' already exists and so hasn\'t been created');
+        }
+
     }
 
     /**
